+++ components/nlus/nlu.ts
@@ -85,6 +85,8 @@ const template: any = {
     }
 };
 
+const errorFileLog = "training.error.log";
+
 export default class Nlu extends Component {
 
     constructor(private helper: IHelper, private api: any) {
@@ -258,8 +260,16 @@ export default class Nlu extends Component {
             console.log("Missing name in nlu.yml");
             return;
         }
+        // check training job
+        const result = await this.helper.toPromise(this.api.nluApi,
+            this.api.nluApi.projectsProjectIdNlusNluNameHasActiveJobGet, projectId, nluName);
+        if (result.data) {
+            console.log("Sorry, your previous training is still running. " +
+                "Give it another try in a few minutes.");
+            return;
+        }
         try {
-            let opts = {};
+            let opts: any = {};
             if (options.file) {
                 console.log(`Training.. (input file: ${options.file})`);
                 opts = {
@@ -272,9 +282,36 @@ export default class Nlu extends Component {
                 };
             }
 
-            const trainResult = await this.helper.toPromise(this.api.nluApi,
+            const { response: { body } } = await this.helper.toPromise(this.api.nluApi,
                 this.api.nluApi.projectsProjectIdNlusNluNameTrainPost, projectId, nluName, opts);
-            console.log(`Success: ${trainResult.data.count} data trained !`);
+            const trainResult = body;
+
+            // Print result
+            const count = trainResult.count;
+            const successCount = trainResult.rowIds ? trainResult.rowIds.length : 0;
+            if (successCount) {
+                console.log(`Success: ${successCount} data trained !`);
+            }
+
+            // Write error to file
+            const errorCount = trainResult.errRows ? trainResult.errRows.length : 0;
+            if (errorCount) {
+                if (options.file) {
+                    const rawData = fs.readFileSync(options.file).toString("utf8");
+                    const trainingData = rawData.split("\n");
+                    let errData = "";
+                    for (const i of trainResult.errRows) {
+                        errData += trainingData[i] + "\n";
+                    }
+                    fs.writeFile(errorFileLog, errData, (err: any) => {
+                        if (err) { throw err; }
+                        console.log(`Error training ${errorCount} data. See details on ${errorFileLog}`);
+                    });
+                } else if (options.sentence) {
+                    console.log(`Error training data`);
+                }
+            }
+
         } catch (error) {
             console.log(this.helper.wrapError(error));
         }
@@ -371,8 +408,9 @@ export default class Nlu extends Component {
     public async snapshot() {
         const projectId = this.helper.getProp("projectId");
         try {
-            const nluDesc : any = this.helper.loadYaml("./nlu.yml");
-            const result = await this.helper.toPromise(this.api.nluApi, this.api.nluApi.projectsProjectIdNluSnapshotGet, projectId, nluDesc.name);
+            const nluDesc: any = this.helper.loadYaml("./nlu.yml");
+            const result = await this.helper.toPromise(this.api.nluApi,
+                this.api.nluApi.projectsProjectIdNlusNluNameSnapshotGet, projectId, nluDesc.name);
 
             console.log(`Snapshot captured!`);
         } catch (error) {